Add My Metronome app
While learning flutter, I'm going to upload a couple of apps to practice flutter and dart.
In this commit, I include an initial structure of a metronome app that's going to be customizable by the user easily.
Every user can change the sound of the metronome and upload his unique sounds.