Introducing a new application FanCate
This project is a mix of a basic social media, an online store and many more.
In a high level, it's a platform to allow individuals to communicate with their fans in a better way.
The can upload their latest projects (image, text, audio, files).
The fans can simply sign-up and support them with a like, comment, share and even a small donation.